+++ modular_sdk/services/impl/maestro_http_transport_service.py
@@ -83,7 +83,9 @@ class MaestroHTTPTransport(AbstractTransport):
         _LOG.debug('Message encrypted')
         # sign headers
         headers = signer.get_signed_headers(
-            async_request=async_request, compressed=compressed,
+            async_request=async_request,
+            compressed=compressed,
+            http=True,
         )
         _LOG.debug('Signed headers prepared')
         return encrypted_body, headers