+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.3.1] - 2024-11-25
+- Update `get_signed_headers` to support conditional handling of `http` and
+`rabbit` specific headers
+
 ## [6.3.0] - 2024-11-20
 - Fix Vault parameters encoder to comply with Maestro format